# docs/api-contract.openapi.yaml
...
openapi: 3.1.0
info:
  title: Parlay chat server API
  version: "2026.08.30"
  description: >
    The /api/chat surface served by both the TS server (packages/server, Bun)
    and the Go server (packages/go-server). No authentication; the surface
    trusts the network boundary (local/tailnet only) plus an origin guard for
    cross-origin browser callers. This file is authoritative over
    docs/api-contract.md on conflict.
servers:
  - url: http://localhost:4242
    description: Default dev/production address (PARLAY_SERVER overrides).
paths:
  /api/chat/send:
    post:
      operationId: send
      summary: Send a user-role message into a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: []
              properties:
                text: { type: string, description: Required unless images present. }
                toAgent: { type: string, description: Target channel (agent id). }
                images:
                  type: array
                  items: { type: string, maxLength: 500 }
                  description: Up to 8 kept; extras/oversize appended to text as plain URLs.
                from: { type: string, description: "Sender attribution, truncated to 40 chars." }
      responses:
        "200":
          description: OkId on success; app-error-on-200 otherwise.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/OkId"
                  - $ref: "#/components/schemas/AppError"
  /api/chat/reply:
    post:
      operationId: reply
      summary: Agent reply on its own channel
      description: >
        TS falls back to environment/session context when `agent` is absent
        and accepts empty text with action/images; Go requires text AND agent.
        May auto-register the agent (response then carries new_channel).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
                agent: { type: string }
                name: { type: string }
                color: { type: string }
                action:
                  $ref: "#/components/schemas/ChatAction"
                  description: TS only.
                images:
                  type: array
                  items: { type: string }
                  description: TS only.
      responses:
        "200":
          description: OkId (+ new_channel iff auto-registered) or app error.
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: "#/components/schemas/OkId"
                      - type: object
                        properties:
                          new_channel: { const: true }
                  - $ref: "#/components/schemas/AppError"
  /api/chat/alert:
    post:
      operationId: alert
      summary: Broadcast a message to multiple channels
      description: >
        Omitted `agents` broadcasts to all — TS includes the global
        (no-channel) history plus every registered agent; Go targets
        registered agents only (x-parlay-divergence 10). An explicit empty
        array delivers to nobody.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                agents:
                  type: array
                  items: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok, channels, delivered]
                    properties:
                      ok: { const: true }
                      channels: { type: integer, description: Channels recorded against. }
                      delivered: { type: integer, description: Live pollers woken. }
                  - $ref: "#/components/schemas/AppError"
...
  /api/chat/message:
    post:
      operationId: postMessage
      x-parlay-server: go-only
      summary: Persist-and-broadcast ingress for out-of-process producers
      description: >
        Stores the message, then broadcasts the resulting `message` SSE event.
        Used by parlay supervise and the PAI hook tailer (hub-ingress.ts).
        The TS server 404s on this path. Status-error convention.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, text]
              properties:
                channel: { type: string }
                role: { type: string, default: agent }
                text: { type: string }
                type: { type: string, description: "ChatMessage.type passthrough, e.g. system_update." }
                source: { type: string }
                meta: { type: object }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkId" }
        "400":
          description: channel and text are required / invalid JSON body.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
...
  /api/chat/history:
    get:
      operationId: history
      summary: Recent chat history, oldest first
      description: >-
        `limit` bounds how many of the retained messages are returned (the
        server retains 5,000). An absent, non-numeric or non-positive `limit`
        returns the whole retained window — there is no smaller server-side
        default. `channel` is accepted by some clients and ignored.
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
          description: Maximum retained messages to return; 0 or absent means all retained.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ChatMessage" }
  /api/chat/poll:
    get:
      operationId: poll
      summary: Agent long-poll (blocks until message or timeout)
      description: >
        Server-side timeout TS 30s / Go 25s (x-parlay-divergence 4). TS
        returns the full ChatMessage; Go a subset with optional
        cursorReset/skipped loud-skip fields (divergences 5-6). Polling an
        unknown non-tombstoned channel auto-registers it on TS. Delivery
        broadcasts message_received. No `after` = wait for next message only.
      parameters:
        - name: after
          in: query
          schema: { type: string }
          description: Last-seen message id (cursor).
        - name: channel
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [timeout]
                    properties:
                      timeout: { const: true }
                  - $ref: "#/components/schemas/ChatMessage"
                  - type: object
                    description: Go poll subset, possibly with loud-skip fields.
                    required: [id, role, text]
                    properties:
                      id: { type: string }
                      role: { type: string }
                      text: { type: string }
                      from: { type: string }
                      cursorReset:
                        const: true
                        description: after-cursor unresolvable; resumed from newest min(50, retained).
                      skipped:
                        type: integer
                        description: Older retained messages left outside the replay window.
        "410":
          description: Channel is tombstoned (unregistered via agent-down).
          content:
            application/json:
              schema:
                type: object
                required: [error, gone]
                properties:
                  error: { type: string }
                  gone: { const: true }
  /api/chat/register-agent:
    post:
      operationId: registerAgent
      summary: Upsert an agent registry entry (idempotent; clears tombstones)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
                name: { type: string }
                color: { type: string }
                nicknames:
                  type: array
                  items: { type: string }
                  description: Explicit [] clears; omitted keeps existing.
                urls: { type: array, items: { type: string } }
                path: { type: array, items: { type: string } }
                caps: { description: "Free-form agent metadata (unrelated to SSE ?caps=)." }
      responses:
        "200":
          description: >
            TS echoes the stored entry ({ok, ...AgentInfo}); Go returns only
            {ok, nicknames?} (x-parlay-divergence 8).
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok]
                    properties:
                      ok: { const: true }
                      id: { type: string }
                      name: { type: string }
                      color: { type: string }
                      nicknames: { type: array, items: { type: string } }
                      urls: { type: array, items: { type: string } }
                      path: { type: array, items: { type: string } }
                  - $ref: "#/components/schemas/AppError"
  /api/chat/unregister:
    post:
      operationId: unregister
      summary: Deregister an agent channel (status-error convention)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
      responses:
        "200":
          description: "{ok, id} on both (divergence 9 converged). Broadcasts agent_unregister."
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { const: true }
                  id: { type: string }
        "400":
          description: id required.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
        "404":
          description: Unknown id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
...
  /api/chat/eval:
    post:
      operationId: eval
      summary: Relay input evaluation to the eval engine; broadcasts input_action
      description: >
        The synchronous response is informational; the device-scoped
        input_action SSE event is the source of truth for applying actions.
        Engine at PARLAY_EVAL_ENGINE_URL (default http://127.0.0.1:4343).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device]
              properties:
                streamId: { type: string, description: Defaults to eval-<device>-main. }
                version: { type: integer }
                text: { type: string }
                cursor:
                  type: object
                  properties:
                    anchor: { type: integer }
                    active: { type: integer }
                reason: { type: string, default: input }
                voiceEnabled: { type: boolean }
                device: { type: string }
                tabs:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: string }
                      name: { type: string }
                      nicknames: { type: array, items: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok, sseClients]
                    properties:
                      ok: { const: true }
                      sseClients: { type: integer }
                      v: { type: integer }
                      streamId: { type: string }
                      seq: { type: integer }
                      baseVersion: { type: integer }
                      actions: { type: array }
                      engineEvalNs: { type: integer }
                      timing: { $ref: "#/components/schemas/EvalTiming" }
                  - $ref: "#/components/schemas/AppError"
        "502":
          description: Engine unreachable or invalid engine response.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AppError" }
  /api/chat/eval-push:
    post:
      operationId: evalPush
      summary: Engine-owned submit fire; broadcasts input_action (serverOwnedFire)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [streamId]
              properties:
                streamId: { type: string }
                seq: { type: integer }
                baseVersion: { type: integer }
                v: { type: integer, description: Defaults to 1. }
                action: { description: Single action; wrapped into the actions array. }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [ok, sseClients]
                properties:
                  ok: { const: true }
                  sseClients: { type: integer }
        "400":
          description: streamId required.
          content: { application/json: { schema: { $ref: "#/components/schemas/AppError" } } }
        "404":
          description: Unknown stream (mapping evicted or engine restarted; client re-registers on next keystroke).
          content: { application/json: { schema: { $ref: "#/components/schemas/AppError" } } }
...
  /api/chat/events:
    get:
      operationId: sseEvents
      summary: The SSE stream (one EventSource per tab)
      description: >
        Connect burst in order — TS: connected, history, agents,
        agent_presence, presence_map; Go adds commands. Keepalive comment
        every 25s. Event names and payloads: see docs/api-contract.md § SSE
        Events (the x-sse-events list below enumerates the names). Guarded on
        Go, unguarded on TS (x-parlay-divergence 28).
      x-sse-events:
        - connected
        - history
        - agents
        - agent_register
        - agent_unregister
        - presence_map
        - message
        - message_received
        - presence
        - draft
        - agent_presence
        - tool_event
        - tts_event
        - lavish_session
        - reload
        - navigate
        - input_action
        - remote_input_result
        - device_cmd
        - pages_patch
        - cursorless_rpc
        - commands
        - command_update
      parameters:
        - name: device
          in: query
          schema: { type: string }
          description: Client-generated uuid; enables device-scoped delivery. TS stores it; Go accepts without storing.
        - name: after
          in: query
          schema: { type: string }
          description: Last-seen message id; resolvable = delta history, else per-channel windowing (TS).
        - name: url
          in: query
          schema: { type: string }
          description: Current page URL; the owning channel's history window widens to 200 (TS).
        - name: caps
          in: query
          schema: { type: string }
          description: >
            Both servers. URL-encoded JSON capability declaration
            (docs/interface-capabilities.md). Invalid = 400, not legacy
            fallback. Gates delivery of navigate/reload/device_cmd/
            input_action/draft, subtraction-only.
      responses:
        "200":
          description: text/event-stream.
          content:
            text/event-stream:
              schema: { type: string }
        "400":
          description: Invalid ?caps= declaration.
          content: { application/json: { schema: { $ref: "#/components/schemas/AppError" } } }
    post:
      operationId: eventsIngress
      x-parlay-server: go-only
      summary: External-producer ingress into the SSE hub (allowlist)
      description: >
        One name per real producer — tool_event alone today. The allowlist is
        derived at init from enrolled source contracts (contracts/sources/,
        observability trust posture); widening it means landing a reviewed
        contract. Everything else is 400, including server-produced names and
        panel-aiming names (those refusal rosters stay hard-coded in the
        handler). Contract owned by events_ingress.go's doc comment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event]
              properties:
                event: { type: string, description: Must be in the allowlist (tool_event). }
                data: { type: object, description: "Forwarded byte-identical; absent broadcasts {}." }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [ok, event]
                properties:
                  ok: { const: true }
                  event: { type: string }
        "400":
          description: event is required / event not accepted from an external producer.
          content: { application/json: { schema: { $ref: "#/components/schemas/AppError" } } }
...
components:
  schemas:
    AppError:
      type: object
      required: [error]
      properties:
        error: { type: string }
      description: >
        The {error} body. Carried on HTTP 200 by the app-error-on-200 routes
        and on non-2xx by the status-error routes.
    OkId:
      type: object
      required: [ok, id]
      properties:
        ok: { const: true }
        id: { type: string, description: Server-assigned message id. }
    ChatAction:
      type: object
      required: [kind, label]
      properties:
        kind: { enum: [navigate, switch_tab] }
        url: { type: string }
        channel: { type: string }
        label: { type: string }
    ChatMessage:
      type: object
      required: [id, role, ts, text]
      properties:
        id: { type: string }
        role: { enum: [user, agent] }
        ts: { type: string, description: ISO 8601. }
        text: { type: string }
        channel: { type: string, description: Agent id; absent for global/user messages. }
        type: { enum: [alert, action_request, system_update] }
        action: { $ref: "#/components/schemas/ChatAction" }
        source: { type: string, description: "system_update: emitting hook/system label." }
        meta: { type: object }
        images: { type: array, items: { type: string } }
        from: { type: string, description: User-role sender attribution; absent = the captain. }
        received: { type: boolean, description: Runtime-only delivery pip; stripped from disk. }
...
    EvalTiming:
      type: object
      properties:
        engineEvalNs: { type: integer }
        relayMs: { type: integer }
        serverOwnedFire: { type: boolean }
...